backend/requirements.txt
flask~=2.0.3
flask_cors~=3.0.10
flask-marshmallow~=0.14.0
flask-sqlalchemy~=2.5.1
marshmallow-sqlalchemy~=0.28.0
SQLAlchemy~=1.4.32
pytest-mock~=3.7.0
pytest-cov~=3.0.0
pytest~=7.1.1
pyyaml~=5.4.1
marshmallow-dataclass~=8.5.3
flask-socketio~=5.1.2
sqlalchemy-mixins~=1.5.0
